[Ionut] - More platform independent code cleanup: -- Moved platform specific defines to their respective header files (e.g. THREAD_LOCAL) -- Moved most preprocessor defines from the VisualC++ projects into source code -- Removed compiler specific code (e.g. warning disables) and fixed most warning resulting from this
git-svn-id: https://xp-dev.com/svn/Divide-Framework/trunk@444 e89c6373-eeba-4702-abef-d26284fe765c

Former-commit-id: 83d55177b3852331d4d9ff75215e7b97df24e0f5 [formerly d9871e3b56b8dd7a2aef5cad7b1997dd81d29434] [formerly b2b10b4fe21d480af254edb568e0d3cd7e3691cb [formerly 700e0e6f6490bf2e5060302792c53d279cc5ffc2]]
Former-commit-id: 987e9f0c89ab40398d9564e2e8df9ec48907228e [formerly 10d318c4fba6ca4e10786b628b3690975b762aa1]
Former-commit-id: 6b2e7f72db324392f91aba4c5fd5deb2217d3874

+++ Source Code/Platform/Video/OpenGL/Headers/GLWrapper.h
@@ -192,9 +192,9 @@ DEFINE_SINGLETON_EXT1_W_SPECIFIER(GL_API, RenderAPIWrapper, final)
     /// is used
     static void togglePrimitiveRestart(bool state);
     /// Set the currently active texture unit
-    static bool setActiveTextureUnit(GLuint unit);
+    static bool setActiveTextureUnit(GLushort unit);
     /// Set the currently active texture unit
-    static bool setActiveTextureUnit(GLuint unit, GLuint& previousUnit);
+    static bool setActiveTextureUnit(GLushort unit, GLuint& previousUnit);
     /// Switch the currently active vertex array object
     static bool setActiveVAO(GLuint ID);
     /// Switch the currently active vertex array object
@@ -236,20 +236,20 @@ DEFINE_SINGLETON_EXT1_W_SPECIFIER(GL_API, RenderAPIWrapper, final)
     /// Bind a texture specified by a GL handle and GL type to the specified
     /// unit
     /// using the sampler object defined by hash value
-    static bool bindTexture(GLuint unit, GLuint handle, GLenum target,
+    static bool bindTexture(GLushort unit, GLuint handle, GLenum target,
                             size_t samplerHash = 0);
     /// Bind multiple textures specified by an array of handles and an offset
     /// unit
-    static bool bindTextures(GLuint unitOffset, GLuint textureCount,
+    static bool bindTextures(GLushort unitOffset, GLuint textureCount,
                              GLuint* textureHandles, GLenum* targets,
                              GLuint* samplerHandles);
 
     /// Bind the sampler object described by the hash value to the specified
     /// unit
-    static bool bindSampler(GLuint unit, size_t samplerHash);
+    static bool bindSampler(GLushort unit, size_t samplerHash);
     /// Bind multiple samplers described by the array of hash values to the
     /// consecutive texture units starting from the specified offset
-    static bool bindSamplers(GLuint unitOffset, GLuint samplerCount,
+    static bool bindSamplers(GLushort unitOffset, GLuint samplerCount,
                              GLuint* samplerHandles);
     /// Return the OpenGL sampler object's handle for the given hash value
     static GLuint getSamplerHandle(size_t samplerHash);
